[COLAB-2191] implemented central post link handler
Slightly hacky implementation:
If a link with the “js-PostLink” class is clicked, cancel the default action (GET this.href). Instead, create a <form> with method=“POST” and action=this.href, add this form to the body, and submit the form.
I chose this implementation over one with $.post(this.href) because it handles redirections as expected. With $.post(this.href), redirections can’t be caught on the client side and would require changes in the controllers on the server side.